% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdsi.R
\docType{package}
\name{nrsmisc}
\alias{nrsmisc}
\alias{nrsmisc-package}
\title{\code{pdsi} package}
\description{
Bare bones package to access and plot Palmer Drought Severity Index data from the
 National Climatic Data Center, available at
 \url{https://www1.ncdc.noaa.gov/pub/data/cirs/climdiv/}
}
