% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdsi_get.R
\name{pdsi_get}
\alias{pdsi_get}
\title{Get US Palmer Drought Severity Index Data}
\usage{
pdsi_get(url = "climdiv-pdsidv-v1.0.0-20180604", update = FALSE)
}
\arguments{
\item{url}{name of current NCDC Palmer Drought Severity Index file
located at \url{https://www1.ncdc.noaa.gov/pub/data/cirs/climdiv/}
(e.g., file in June 2018 is 'climdiv-pdsidv-v1.0.0-20180604')}

\item{update}{logical (default \code{FALSE}); update current PDSI data
associated with package?}
}
\description{
If the package already has the most current version of the data, you can
simply load the \code{pdsi} data set using \code{data(pdsi)}
}
\examples{
\dontrun{
# File name changes with each NCDC monthly update, so modify accordingly
get_pdsi("climdiv-pdsidv-v1.0.0-20180604")
}
}
