% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pdsi.R
\name{pdsi_plot}
\alias{pdsi_plot}
\title{Generate interactive plot US Palmer Drought Severity Index Data}
\usage{
pdsi_plot(lat = NULL, lon = NULL, address = NULL, nyrs = 10)
}
\arguments{
\item{lat}{numeric scalar of position latitude (decimal degrees; WGS84)}

\item{lon}{numeric scalar of position longitude (decimal degrees; WGS84)}

\item{address}{character scalar of a street address or place name
(e.g. "Mattamuskeet NWR" or "135 Phoenix Rd, Athens, GA"); overrides
\code{lat} and \code{lon} if specified. See example}

\item{nyrs}{integer (default = 10); how many years before present would
you like the initial graph to display?}
}
\description{
Generate interactive plot US Palmer Drought Severity Index Data
}
\examples{
\dontrun{
plot_pdsi(34, -83)
plot_pdsi(address = "Alligator River NWR")
# Return only ASOS stations...
find_wx_stns(address = "Mattamuskeet NWR", asos_only = TRUE, plot = TRUE)
}
}
